chore(e2e): move calibration honesty cluster into its own spec
Cuts the 9-test `calibration honesty layer` describe block out of
forecast-regional.spec.ts (where it was misfiled — the assertions hit
beach-detail URLs, not /forecast/[region]) into its own
calibration-honesty.spec.ts.

Loosens the 3 calibrated-beach assertions to drop a wave-size-sensitive
not.toContain("~") check. The unified-surf-card hero uses per-reading
isCalibrated (source === 'cdip_sig' && shoaling_factors), so on tiny-wave
days at calibrated beaches the hero correctly downgrades to "Forecast
height" + tilde while forecast-tab (beach-level) stays "Face height" —
the existing getByText(/Face height/i) check is the real beach-level
contract worth pinning.

Extends e2e/utils/error-detection.ts:isPlaceholderImageProxyFailure to
also ignore status 424 on /api/image-proxy. Same graceful-degradation
class as 400/403, just a different upstream code (Openverse returned
424 on La Jolla Shores). Matches the existing comment intent: "ignore
ALL image proxy failures since they don't affect core functionality."

Verified against dev.quiversurf.app: calibration-honesty.spec.ts 9/9
pass, forecast-regional.spec.ts 22/22 pass.

Resolves the calibration cluster in issue #231 (~11 of 100 failures).
The companion quick win there — deleting forecast-om-primary.spec.ts —
already shipped in 851df440.

Plan: ~/.claude/plans/yes-clever-hartmanis.md
